Fixed pylint
